requirements.txt
Django==5.2.8
python-telegram-bot==21.5
python-dotenv==1.0.0
google-generativeai==0.7.2
requests==2.31.0
Pillow==10.1.0
pandas==2.1.3
matplotlib==3.8.2
APScheduler==3.10.4
pytz==2023.3
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-dateutil==2.8.2
asgiref==3.7.1
